requirements.txt
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
graphviz==0.20.3
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pydot==2.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
threadpoolctl==3.4.0
tqdm==4.66.2
